Add checks for student creation
Added checks for student creation for:
A) If the user tries to create a new student with an ID that is already in
the database, the program no longer crashes.
B) If the user tries to enter a student ID which contains non-numeric
characters, the program no longer crashes.
C) If the user tries to enter a student graduation year which contains
non-numeric characters, the program no longer crashes.